second commit

+++ program1.c
@@ -56,7 +56,7 @@ int promColor(int rgb_val){
     int red = (rgb_val & 0xFF0000) >> 16;
     int green = (rgb_val & 0x00FF00) >> 8;
     int blue = rgb_val & 0x0000FF;
-    int prominent = fmax(fmax(red, green), blue);
+    int prominent = fmaxf(fmaxf(red, green), blue);
     
     if (prominent == red & red != green & red != blue) return 1;
     else if (prominent == green & green != blue & green != red) return 2;
